Feature: Implementing deletion of multiple problems on the Problems page for instructors. THis time it checks for student data to make sure none exists.

+++ Instructors/views/challengeQuestionsListView.py
@@ -1,12 +1,13 @@
 from django.template import RequestContext
 from django.shortcuts import render,redirect
 from django.contrib.auth.decorators import login_required, user_passes_test
-from Instructors.models import Challenges, Courses, CoursesSkills, ChallengesQuestions
+from Instructors.models import Challenges, Courses, CoursesSkills, ChallengesQuestions, Questions
 from Instructors.views.challengeListView import makeContextDictForQuestionsInChallenge
 from Instructors.views.searchResultsView import searchResults
 from Instructors.views.utils import initialContextDict
 from Instructors.lupaQuestion import lupa_available
 from Instructors.constants import unassigned_problems_challenge_name
+from Students.models import StudentChallengeQuestions
 
 from Badges.enums import dict_dict_to_zipped_list
 from Instructors.questionTypes import QuestionTypes
@@ -55,11 +56,6 @@ def challengeQuestionsListView(request):
  
     context_dict, currentCourse = initialContextDict(request)
     context_dict['lupa_available'] = lupa_available
-
-    if request.POST:
-        if 'deletion_checkboxes' in request.POST:
-            PerformDeletion(request.POST.getlist('deletion_checkboxes'))
-            return redirect("/oneUp/instructors/challengeQuestionsList?problems")
         
     if request.GET:
         if 'challengeID' in request.GET:   
@@ -96,8 +92,44 @@ def challengeQuestionsListView(request):
 
     return render(request,'Instructors/ChallengeQuestionsList.html', context_dict)
 
-def PerformDeletion(problems):
+def makeSureTheresNoStudentsUsingThatProblem(problem):
+    isAstudentUsingTheProblem = True
+    print("len(StudentChallengeQuestions.objects.filter(challengeQuestionID=int(problem)))", len(StudentChallengeQuestions.objects.filter(challengeQuestionID=int(problem))))
+    if(len(StudentChallengeQuestions.objects.filter(challengeQuestionID=int(problem))) == 0):
+        isAstudentUsingTheProblem = False
+    return isAstudentUsingTheProblem
+
+def performDeletion(problems):
+    errorList = []
     for problem in problems:
-        question = ChallengesQuestions.objects.get(questionID=int(problem))
-        print("question" + str(question))
-        question.delete()
+        challengeQuestion = ChallengesQuestions.objects.get(questionID=int(problem))
+        if makeSureTheresNoStudentsUsingThatProblem(problem):
+            errorList.append("Error: Student data found for problem: "+ challengeQuestion.questionID.preview)
+            print("data found",  challengeQuestion.questionID.preview)
+        else:
+            question = Questions.objects.get(questionID=int(problem))
+            question.delete()
+            challengeQuestion.delete()
+            print("deleted")
+        
+    return errorList
+@login_required
+def deleteProblemsButFilterTakenByStudent(request):
+    ##this is used to track how many times the student clicks class average
+    ##we use ajax to track the information, otherwise they'd get the page refreshed on them
+    ##and it would be "wrong".
+    from django.http import JsonResponse
+    
+    context_dict, currentCourse = initialContextDict(request)
+    context_dict['lupa_available'] = lupa_available
+
+    errorList = []
+    response = {}
+    
+    if request.POST:
+        if 'deletion_checkboxes' in request.POST:
+            errorList = performDeletion(request.POST.getlist('deletion_checkboxes'))
+            response['errorMessages'] = errorList
+            
+    return JsonResponse(response)
+
